+++ scripts/goodreads-obsidian-base-readlist/src/lib/helpers/getDatabase/getGenres.ts
@@ -1,0 +1,14 @@
+export function getGenres(): string[] {
+  const genresRaw = document.querySelectorAll('.BookPageMetadataSection__genreButton');
+  const genres: string[] = [];
+
+  genresRaw.forEach((item) => {
+    const content = item.querySelector('.Button__labelItem')?.textContent.trim() || '';
+
+    if (content) {
+      genres.push(content);
+    }
+  });
+
+  return genres;
+}