+++ scripts/goodreads-obsidian-base-readlist/src/main.ts
@@ -1,0 +1,10 @@
+import GrabDetails from './lib/GrabDetails/GrabDetails.svelte';
+import { mountSvelteComponents } from '@repo/shared-browser';
+
+mountSvelteComponents([
+  {
+    component: GrabDetails,
+    target: document.querySelector('.BookActions'),
+    styles: { display: 'inline-flex' },
+  },
+]);