+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @repo/shared
 
+## 2.0.2
+
+### Patch Changes
+
+- Fix formatting
+
 ## 2.0.1
 
 ### Patch Changes