+++ scripts/imdb-obsidian-base-watchlist/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 2.3.5
+
+### Patch Changes
+
+- Fix formatting
+- Updated dependencies
+  - @repo/shared-browser@1.5.1
+  - @repo/shared@2.0.2
+
 ## 2.3.4
 
 ### Patch Changes